# backend/src/main/java/com/example/gplx/auth/dto/request/LoginRequest.java
package com.example.gplx.auth.dto.request;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;

public record LoginRequest(
        @Email @NotBlank String email,
        @NotBlank String password
) {
}